requirements.txt
python-dotenv
google-generativeai
livekit-agents
livekit-plugins-silero
sounddevice
soundfile
numpy
gTTS
playsound